refactor(properties drawer): create properties drawer directory for components
