+++ src/components/ui/PropertiesDrawer/OverviewTable.tsx
@@ -1,12 +1,8 @@
 import React from 'react';
-import { File } from '../../shared.types';
-import { formatDate, formatFileSize } from '../../utils';
+import { File } from '../../../shared.types';
+import { formatDate, formatFileSize } from '../../../utils';
 
-type FileOverviewTableProps = {
-  file: File | null;
-};
-
-export default function FileOverviewTable({ file }: FileOverviewTableProps) {
+export default function OverviewTable({ file }: { file: File | null }) {
   return (
     <div className="w-full overflow-hidden rounded-lg border border-surface mt-4">
       <table className="w-full">
@@ -15,7 +11,7 @@ export default function FileOverviewTable({ file }: FileOverviewTableProps) {
             <td className="p-3 border-b border-surface bg-surface-light text-sm text-foreground dark:bg-surface-dark font-medium">
               Last modified
             </td>
-            <td className="p-3 ">
+            <td className="p-3">
               {file ? formatDate(file.last_modified) : null}
             </td>
           </tr>
@@ -23,7 +19,7 @@ export default function FileOverviewTable({ file }: FileOverviewTableProps) {
             <td className="p-3 border-b border-surface bg-surface-light text-sm text-foreground dark:bg-surface-dark font-medium">
               Size
             </td>
-            <td className="p-3 ">
+            <td className="p-3">
               {file ? (file.is_dir ? '—' : formatFileSize(file.size)) : null}
             </td>
           </tr>
